==> scripts/pi_curves/2_PI_Curve_NLS_Fitting.Rmd <==
---
title: "Oxygen flux rate extractions "
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## install packages if you dont already have them in your library
if (!require("devtools")) install.packages("devtools")
if (!require("furrr")) install.packages("furrr")
if (!require("future")) install.packages("future")
if (!require("tidyverse")) install.packages("tidyverse")
if (!require("lubridate")) install.packages("lubridate")
if (!require("cowplot")) install.packages("cowplot")
if (!require("LoLinR")) install_github('colin-olito/LoLinR') 

## load libraries
library(devtools)
library(LoLinR)
library(tidyverse)
library(lubridate)
library(cowplot)
library(broom)
library(plotrix)
library(Hmisc)
library(rTPC)
library(nls.multstart)
library(here)
library(nlraa)
library(ggplot2)
library(ggstatsplot)

## libraries for parallel processing
library(future)
library(furrr)
```

# Define data   

```{r} 
#specify data 
Data<-read_csv(here("output/pi_curves/","pi_curve_extracted_rates.csv"))

Data$PAR <- as.numeric(Data$Light_Value.x) 
Data$Pc <- as.numeric(Data$micromol.cm2.s) 
```

```{r}
Data %>% 
ggplot(aes(x=PAR, y=Pc, color=Species, fill=Species))+ 
  geom_point() 

raw.plot <- Data %>% 
ggplot(aes(x=PAR, y=Pc, color=Species, group=sample_id, fill=Species))+ 
  geom_point()+
  geom_line()+
  facet_wrap(~ sample_id,  ncol = 10)+
  theme(legend.position = "none")+
  theme_classic()

ggsave("figures/pi_curves/pi_curve_raw_plots_bycoral.png",raw.plot,  width = 10, height=6)

raw.plot.2 <- Data %>% 
ggplot(aes(x=PAR, y=Pc, color=Species, group=sample_id, fill=Species))+ 
  geom_point()+
  geom_line()+
  facet_wrap(~ Species * Lifestage,  ncol = 2)+
    theme_classic()+
  theme(legend.position="none")

ggsave("figures/pi_curves/pi_curve_raw_plots_groups.png",raw.plot.2,  width = 8, height=8)

raw.plot.3 <- Data %>% 
ggplot(aes(x=PAR, y=Pc, color=Species, group=sample_id, fill=Species, linetype=Lifestage))+ 
  geom_point()+
  geom_line()+
  facet_wrap(~ Species,  ncol = 3)+
  theme(legend.position = "none")+
  theme_classic()

ggsave("figures/pi_curves/pi_curve_raw_plots_species.png",raw.plot.3,  width = 10, height=6)

ests <- Data%>%
  group_by(Species, Lifestage, Temp.Cat.x, sample_id)%>%
  summarise(Rd= min(Pc),
            Pmax = max(Pc))

#ests$Temp.Meas <-c(11.9, 18.15, 22.2, 25.55, 27.95,  31.8, 34.95, 38.95,
    #               11.9, 18.15, 22.2, 25.55, 27.95,  31.8, 34.95, 38.95,
    #               11.9, 18.15, 22.2, 25.55, 27.95,  31.8, 34.95, 38.95)

dirty.tpc.plot.Rd <- ests %>% 
ggplot(aes(x=Temp.Cat.x, y=-Rd, color=Species, group=Species, fill=Species, shape=Lifestage))+ 
  geom_point()+
  #geom_line()+
  facet_wrap(~ Species * Lifestage, ncol=2)+
  theme_classic()+
  theme(legend.position = "none"); dirty.tpc.plot.Rd  

#ggsave("figures/pi_curves/rd_dirty_estimates.png",dirty.tpc.plot.Rd,  width = 10, height=6)

dirty.tpc.plot.Pmax <- ests %>% 
ggplot(aes(x=Temp.Cat.x, y=Pmax, color=Species, group=Species, fill=Species, shape=Lifestage))+ 
  geom_point()+
  #geom_line()+
  facet_wrap(~ Species * Lifestage, ncol=2)+
  theme(legend.position = "none")+
  theme_classic(); dirty.tpc.plot.Pmax 

#ggsave("figures/pi_curves/pmax_dirty_estimates.png",dirty.tpc.plot.Pmax,  width = 10, height=6)

``` 

# PI Curves 

Define PI curve function as a nonlinear Least Squares regression of a quadratic fit, test nls fit 

Pc ~ (Am*((AQY*PAR)/(sqrt(Am^2 + (AQY*PAR)^2)))-Rd), data=., start=list(Am=(max(.$Pc)-min(.$Pc)),  AQY=0.001, Rd=-min(.$Pc)) 

Aquatic Photosynthesis, Falkowski  
PAR = irradiance from 400-700nm (also called I or E) 
PC = oxygen flux rate 
Pmax = max photosynthesis (also called Am)   
alpha = quantum yeild (also called AQY)   
I or E or PAR = irradiance  
Rd = dark respiration 
Ik (saturating irradiance) is the point at which photosynthesis reaches the max of initial slope = Am/AQY 
Ic=(Am*Rd)/(AQY*(sqrt(Am^2-Rd^2))) 
Equation for Ic derived from quadratic equation above. Ic = Par when Pc = 0 (x intercept). Ic = light compensation point; point at which photosynthesis is released from carbon limitation.  

Run nls model  
Using flexible initial values based on input data: 

Example
```{r}
set.seed(1234)

Data$PAR <- as.numeric(Data$Light_Value.x) 
Data$Pc <- as.numeric(Data$micromol.cm2.s) 

x <- Data$PAR
y <- Data$Pc
dat <- data.frame(x = x, y = y)
fit <- nls(y ~ SSnrh(x, asym, phi, theta, rd), data = dat)
## Visualize observed and simulated
ggplot(data = dat, aes(x = x, y = y)) + 
  geom_point() + 
  geom_line(aes(y = fitted(fit)))

## Testing predict function
prd <- predict_nls(fit, interval = "confidence")
datA <- cbind(dat, prd)

## Plotting
ggplot(data = datA, aes(x = x, y = y)) + 
  geom_point() + 
  geom_line(aes(y = fitted(fit))) + 
  geom_ribbon(aes(ymin = Q2.5, ymax = Q97.5), 
  fill = "purple", alpha = 0.3)
```

Problem corals that don't fit model: 
ACR-A4 - flat on upper part of light levels
POC-A5 - doesn't level off? 
POR-A5 - flat ish curve
ACR-R6 - flat curve
ACR-R8 - not sure on this one
ACR-R9 - spike in lower levels of light
POC-R3 - maybe weird curve
POC-R4 - doesn't flatten out
POC-R6 - doesn't flatten out
POC-R9 - weird curve maybe
POC-R10 - peak in lower light
POR-R4 - weird curve maybe
POR-R5 - doesn't reach flattened part of curve
POR-R6 - doesn't flatten out 
POR-R7 - doesn't flatten out
POR-R9 - doesn't flatten out 
POR-R10 - doesn't flatten out 

Run without these
```{r} 
nls_data <- Data %>%  
  filter(sample_id %in% c("ACR-A1", "ACR-A2", "ACR-A3", "ACR-A5", "POC-A1", "POC-A2", "POC-A3", "POC-A4", "POR-A1", "POR-A2", "POR-A3", "POR-A4", "ACR-R1", "ACR-R2", "ACR-R3", "ACR-R4", "ACR-R5", "ACR-R7", "ACR-R10", "POC-R1", "POC-R2","POC-R5", "POC-R7", "POC-R8", "POR-R1", "POR-R2", "POR-R3", "POR-R8"))%>%
  
  nest(.by = sample_id)%>%
  mutate(model1 = map(data, ~  
                        #nls(Pc ~ (Am*((AQY*PAR)/(sqrt(Am^2 + (AQY*PAR)^2)))-Rd), data=., start=list(Am=0.001,  AQY=0.001, Rd=-0.05)) %>%
                         nls(Pc ~ SSnrh(PAR, asym, phi, theta, rd), data = .)%>% 
#asym=asymptote, 
#phi=quantum efficiency, 
#theta = curvature 
#rd= dark respiration                  
                              tidy %>% 
                              dplyr::select(term, estimate) %>%  
                              spread(term, estimate))) %>% 
  unnest(model1) %>% 
  unnest(c(data)) %>% 
  group_by(sample_id) %>% 
  summarise(Am=mean(asym), AQY=mean(phi), theta=mean(theta), Rd=mean(rd))%>% 
  mutate(Ik=Am/AQY)%>% 
  mutate(Ic=(Am*Rd)/(AQY*(sqrt(Am^2-Rd^2)))) 

write_csv(nls_data, "output/pi_curves/pi_curve_pars_nls.csv") 
``` 


Plot curve over data points. NOT WORKING   
```{r} 
augmented <- Data %>%  
  nest(.by = sample_id)%>%  
  mutate( 
    fit = map(data, ~ nls(Pc ~ (Am*((AQY*PAR)/(sqrt(Am^2 + (AQY*PAR)^2)))-Rd), data=., start=list(Am=0.7,  AQY=0.001, Rd=.4))), 
    augmented = map(fit, augment), 
  ) %>%  
  unnest(augmented) 

#all colonies together 
augmented %>% 
  group_by(sample_id)%>% 
  qplot(PAR, Pc, data = ., geom = 'point', colour = sample_id) + 
  geom_line(aes(y=.fitted))+ 
  theme(legend.position="none") 

#view by certain colonies to ground truth values  
augmented %>% 
  group_by(sample_id)%>% 
  #filter(sample_id=="ACR-139")%>% 
  qplot(PAR, Pc, data = ., geom = 'point', colour = sample_id) + 
  geom_line(aes(y=.fitted))+ 
  theme(legend.position="none") 

#view individual plots 
by(augmented,augmented$sample_id, function(i) { 
  ggplot(i) + 
          geom_point(aes(PAR, Pc, group=sample_id)) +  
          geom_line(aes(y=.fitted, x=PAR)) +  
          theme_classic()+ 
          labs(x = expression(paste('PAR (', mu, "mol photons m"^-2, 's'^-1,")")), 
               y = expression(paste('Photosynthetic rate (', mu, "mol cm"^-2, 'h'^-1,")")), 
               title = paste0("1_", augmented$sample_id)) 
}) 

``` 

PI curve fitted parameters for each individual 
```{r} 
pars <- nls_data 
md <- read_csv("data/pi_curves/pi_curves_sample_metadata.csv") %>% select(sample_id, Species, Lifestage)
df <- left_join(pars, md) 

df <- df %>% 
  pivot_longer(cols=Am:Ik, names_to="variable") 

# Facet grid for each variable 
PI.params <- df %>% 
  filter(value>-5)%>% #remove outlier
  ggplot(aes(x = Species, y = value, colour=Lifestage)) + 
  facet_wrap(~variable, scales="free_y")+
  geom_point(position=position_jitterdodge(0.1))+
  #scale_colour_manual(values=c("black", "orange"))+ 
  theme_classic()+ 
  theme(legend.position="right");PI.params 

#ggsave("figures/pi_curves/PI_parameters_untransformed.jpeg", PI.params, dpi=300, w=11, h=8, units="in") 
``` 

Load rate data and metadata 
```{r, warning=FALSE, message=FALSE} 
pars <- read_csv("output/pi_curves/pi_curve_pars_nls.csv") 
md <- read_csv("data/pi_curves/pi_curves_sample_metadata.csv") 
df <- left_join(pars, md) 

df$Am.log10 <-log10(df$Am+1) 
df$AQY.log10 <-log10(df$AQY+1) 
df$Rd.log10 <-log10(df$Rd+1) 
df$Ik.log10 <-log10(df$Ik+1) 
df$Ic.log10 <-log10(df$Ic+1)

# add a group name by pasting Temperature and Treatment 
df$group <- paste0(df$Species,"_", df$Lifestage) 

df <- df %>% 
  select(sample_id, Species, Lifestage, Am.log10:Ic.log10)%>%
  pivot_longer(cols=Am.log10:Ic.log10, names_to="variable") 
``` 

Visualize data transformed  
```{r, warning=FALSE, message=FALSE} 
# Visualize data 
values_plot_transformed<-df %>% 
    filter(value>-5)%>%
    ggplot(., aes(x = Species, y = value, colour=Lifestage, fill=Lifestage)) + 
    geom_point(aes(fill=Lifestage, group=Lifestage), pch = 19, size=2, alpha=1, position=position_dodge(0.3)) +  
    xlab("Species") +  
    #scale_fill_manual(values=c("black", "orange"))+
   #scale_colour_manual(values=c("black", "orange"))+
    facet_wrap(vars(variable), scales = "free_y", nrow = 2, strip.position = "top") + 
    theme_classic() +  
    theme(legend.position="right", 
      axis.title=element_text(face="bold", size=16), 
      axis.text=element_text(size=12, color="black"),  
      legend.title=element_text(face="bold", size=14),  
      legend.text=element_text(size=12)); values_plot_transformed 

ggsave("figures/pi_curves/PI_parameters_transformed.jpeg", values_plot_transformed, dpi=300, w=11, h=8, units="in") 
``` 

# ANOVA tests 

Test for differences in parameters by group. 

Am
```{r}
model1<-df%>%
  filter(variable=="Am.log10")%>%
  
  aov(value ~ Species * Lifestage, data=.)

summary(model1)
```
Am different by species and lifestage. 


LEDR
```{r}
model2<-df%>%
  filter(variable=="Rd.log10")%>%
  
  aov(value ~ Species * Lifestage, data=.)

summary(model2)
```
LEDR different by species and lifestage. 

AQY
```{r}
model3<-df%>%
  filter(variable=="AQY.log10")%>%
  
  aov(value ~ Species * Lifestage, data=.)

summary(model3)
```
AQY different by species. 

Ik
```{r}
model4<-df%>%
  filter(variable=="Ik.log10")%>%
  
  aov(value ~ Species * Lifestage, data=.)

summary(model4)
```
Ik different by lifestage. 

Ic
```{r}
model5<-df%>%
  filter(variable=="Ic.log10")%>%
  
  aov(value ~ Species * Lifestage, data=.)

summary(model5)
```
Ic different by lifestage.
